% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{ace_cor}
\alias{ace_cor}
\title{Calculates ace based transformations and correlation, handling missing values and factors.}
\usage{
ace_cor(x, y, handle.na = TRUE)
}
\arguments{
\item{x}{a numeric vector or factor}

\item{y}{a numeric vector or factor}

\item{handle.na}{If TRUE uses pairwise complete observations.}
}
\value{
result of acepack::ace
}
\description{
Calculates ace based transformations and correlation, handling missing values and factors.
}
\examples{
ace_cor(iris$Sepal.Length, iris$Species)
}
