% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{as.matrix.pairwise}
\alias{as.matrix.pairwise}
\title{Converts a pairwise to a symmetric matrix. Uses the first entry for each (x,y) pair.}
\usage{
\method{as.matrix}{pairwise}(x, ...)
}
\arguments{
\item{x}{An object of class pairwise}

\item{...}{other arguments}
}
\value{
A symmetric matrix
}
\description{
Converts a pairwise to a symmetric matrix. Uses the first entry for each (x,y) pair.
}
