% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullseye-package.R
\docType{package}
\name{bullseye-package}
\alias{bullseye}
\alias{bullseye-package}
\title{bullseye: Visualising multiple pairwise variable correlations and other scores}
\description{
We provide a tidy data structure and visualisations for multiple or grouped variable correlations, general association measures and other pairwise scores suitable for numerical, ordinal and nominal variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cbhurley.github.io/bullseye/}
}

}
\author{
\strong{Maintainer}: Catherine Hurley \email{catherine.hurley@mu.ie}

Authors:
\itemize{
  \item Amit Chinwan \email{amit.chinwan.2019@mumail.ie}
}

}
\keyword{internal}
