% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_db.R
\name{download_and_unzip}
\alias{download_and_unzip}
\title{Download and unzip a file}
\usage{
download_and_unzip(url, temp_dir)
}
\arguments{
\item{url}{The URL of the ZIP file to download.}

\item{temp_dir}{The directory where the ZIP file will be downloaded and unzipped.}
}
\value{
The directory where the unzipped files are stored.
}
\description{
Download and unzip a file
}
