% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_db.R
\name{fetch_zip_links}
\alias{fetch_zip_links}
\title{Fetch ZIP file links from a given URL}
\usage{
fetch_zip_links(base_url)
}
\arguments{
\item{base_url}{The base URL from which to fetch ZIP links.}
}
\value{
A character vector containing the full URLs of the ZIP files.
}
\description{
Fetch ZIP file links from a given URL
}
