% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_db.R
\name{main_pipeline}
\alias{main_pipeline}
\title{Main pipeline function for fetching, downloading, and reading shapefiles}
\usage{
main_pipeline(
  remove_temp = FALSE,
  base_url,
  filename,
  compression = "zstd",
  compression_level = 6,
  n_limit = NULL,
  save = TRUE
)
}
\arguments{
\item{remove_temp}{Logical; should the temporary directory be removed? Default is FALSE.}

\item{base_url}{The base URL from which to fetch ZIP links.}

\item{filename}{The filename for the output GeoParquet file in inst/extdata.}

\item{compression}{The compression algorithm for GeoParquet. Default is "zstd".}

\item{compression_level}{The compression level. Default is 6.}

\item{n_limit}{Optional; limit the number of ZIP files to download. Default is NULL (download all).}
}
\value{
The output is written as a GeoParquet file.
}
\description{
Main pipeline function for fetching, downloading, and reading shapefiles
}
