% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pair_all}
\alias{pair_all}
\title{Calculates multiple association measures}
\usage{
pair_all(
  d,
  scores = c("pair_cor", "pair_dcor", "pair_mine", "pair_ace", "pair_cancor", "pair_nmi",
    "pair_polycor", "pair_tau", "pair_gkGamma", "pair_gkTau", "pair_uncertainty",
    "pair_chi"),
  handle.na = T
)
}
\arguments{
\item{d}{dataframe}

\item{scores}{a vector of names of pairwise functions.}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate pairwise score, otherwise NAs not handled.}
}
\value{
tibble of class "multi_pairwise"
}
\description{
Calculates multiple association measures for every variable pair in a dataset.
}
\examples{
iris1 <- iris
iris1$Sepal.Length <- cut(iris1$Sepal.Length,3)
pair_all(iris1)
}
