% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_cancor}
\alias{pair_cancor}
\title{Canonical correlation}
\usage{
pair_cancor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate correlation coefficient,, otherwise NAs not handled.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with canonical correlation for every numeric or factor or mixed variable pair
}
\description{
Calculates canonical correlation for every variable pair in a dataset.
}
\examples{
pair_cancor(iris)
}
