% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_scores.R
\name{pair_control}
\alias{pair_control}
\title{Default scores calculated  by \code{pair_scores}}
\usage{
pair_control(
  nn = c("pair_cor", "pair_dcor", "pair_mine", "pair_ace", "pair_cancor", "pair_nmi",
    "pair_scagnostics"),
  fn = c("pair_cancor", "pair_nmi", "pair_ace"),
  ff = c("pair_cancor", "pair_ace", "pair_nmi", "pair_polycor", "pair_tau",
    "pair_gkGamma", "pair_gkTau", "pair_uncertainty", "pair_chi"),
  nnargs = NULL,
  fnargs = NULL,
  ffargs = NULL
)
}
\arguments{
\item{nn}{function for numeric pairs of variables, should return object of class \code{pairwise}. Use NULL to ignore numeric pairs.}

\item{fn}{function for factor-numeric pairs of variables, should return object of class \code{pairwise}. Use NULL to ignore factor-numeric pairs.}

\item{ff}{function for factor pairs of variables, should return object of class \code{pairwise}. Use NULL to ignore factor-factor pairs.}

\item{nnargs}{other arguments for the nn function}

\item{fnargs}{other arguments for the fn function}

\item{ffargs}{other arguments for the ff function}
}
\value{
list
}
\description{
Gives a list specifying the function to be used for two numeric (nn) variables, two factors (ff)
and for a factor-numeric pair (fn).
}
