% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_dcor}
\alias{pair_dcor}
\title{Distance correlation}
\usage{
pair_dcor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate distance correlation, otherwise NAs not handled.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with distance correlation for every numeric variable pair,
or NULL if there are not at least two numeric variables
}
\description{
Calculates distance correlation for every numeric variable pair in a dataset.
}
\details{
The distance correlation is calculated using \code{\link[energy]{dcor2d}} from \code{energy} package
}
\examples{
pair_dcor(iris)
}
