% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\docType{data}
\name{pair_methods}
\alias{pair_methods}
\title{Pairwise score functions available in the package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 7 columns.
}
\usage{
pair_methods
}
\value{
tibble
}
\description{
A tibble of score functions along with the types of variable pairs these functions
can be applied to. It also contains information regarding the packages used to calculate scores
and the range of the values calculated.
}
\examples{
 pair_methods
}
\keyword{datasets}
