% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_polychor}
\alias{pair_polychor}
\title{Polychoric correlation}
\usage{
pair_polychor(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{ignored. Pairwise complete observations are used automatically.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with polychoric correlation for factor pairs,
or NULL if there are not at least two factor variables
}
\description{
Calculates Polychoric correlation using  for every factor variable pair in a dataset.
}
\details{
The polychoric correlation is calculated using the \code{\link[polycor]{polychor}} function from the
\code{polycor} package, and assumes factor levels are in the given order. NAs are automatically handled by pairwise omit.
}
\examples{
pair_polychor(iris)
}
