% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_scores.R
\name{pair_scores}
\alias{pair_scores}
\title{Calculates association or conditional association measures for a dataset}
\usage{
pair_scores(
  d,
  by = NULL,
  ungrouped = TRUE,
  control = pair_control(),
  handle.na = TRUE
)
}
\arguments{
\item{d}{a dataframe}

\item{by}{a character string for the name of the conditioning variable. Set to \code{NULL} by default.}

\item{ungrouped}{Ignored if \code{by} is \code{NULL}.
If TRUE calculates the overall measure of association for every pair of
variable, in addition to association measures for pairs at levels of the
conditioning variable.}

\item{control}{a list for the measures to be calculated for different variable types. The default is
\code{\link[=pair_control]{pair_control()}} which calculates Pearson's correlation if the variable pair is numeric,
and canonical correlation for factor or mixed pairs.}

\item{handle.na}{If TRUE uses pairwise complete observations to calculate measure of association.}
}
\value{
A tibble with class \code{pairwise}.
}
\description{
Calculates association measures for every variable pair in a dataset when \code{by} is \code{NULL}. If \code{by}
is a name of a variable in the dataset, conditional association measures for every
variable pair at different levels of the grouping variable are calculated.
}
\details{
Returns a \code{pairwise} tibble structure.
}
\examples{
irisc <- pair_scores(iris)
irisc <- pair_scores(iris, control=pair_control(nnargs= list(method="spearman")))
irisc <- pair_scores(iris, control=pair_control(fn="pair_ace"))

#Lots of numerical measures
irisc <- pair_scores(iris, control=pair_control(nn="pair_all", fn=NULL))
irisc <- pair_scores(iris, control=pair_control(nn=c("pair_all"), nnargs=c("pair_cor"), fn=NULL))
#conditional measures
cond_iris <- pair_scores(iris, by = "Species") 
cond_iris_wo <- pair_scores(iris, by = "Species",ungrouped=FALSE) # without overall
irisc <- pair_scores(iris, control=pair_control(nn="pair_all", fn=NULL))
irisc <- pair_scores(iris, by = "Species",control=pair_control(nn="pair_all", fn=NULL))

#scagnostics
sc <- pair_scores(iris, control=pair_control(nn="pair_scagnostics", fn=NULL)) # ignore fn pairs
sc <- pair_scores(iris, by = "Species",
                  control=pair_control(nn="pair_scagnostics", fn=NULL)) # ignore fn pairs
}
