% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_tau}
\alias{pair_tau}
\title{Kendall's tau A, B, C and Kendall's W}
\usage{
pair_tau(d, method = c("B", "A", "C", "W"), ...)
}
\arguments{
\item{d}{A dataframe}

\item{method}{A character string for the correlation coefficient to be calculated, one of "B" (default),
"A", "C" or "W". If the value is "all", then all four correlations are calculated.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with factor pairs along with one of either Kendall's tau A, B, C or
Kendall's W value, or NULL if there are not at least two factor variables
}
\description{
Calculates one of either Kendall's tau A, B, C or Kendall's W for every  factor variable pair in a dataset.
}
\details{
The association values Kendall's tau A, B, C or Kendall's W are calculated using \code{\link[DescTools]{KendallTauA}},
\code{\link[DescTools]{KendallTauB}}, \code{\link[DescTools]{StuartTauC}} or \code{\link[DescTools]{KendallW}} respectively,from the
\code{DescTools} package, and assumes factor levels are in the given order. NAs are automatically handled by pairwise omit.
}
\examples{
 pair_tau(iris)
 pair_tau(iris, method="A")
 pair_tau(iris, method="C")
 pair_tau(iris, method="W")
}
