% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_tauB}
\alias{pair_tauB}
\title{Kendall's tau  B for association between ordinal factors.}
\usage{
pair_tauB(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{ignored. Pairwise complete observations are used automatically.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with factor pairs, or NULL if there are not at least two factor variables
}
\description{
Calculates Kendall's tau B every  factor variable pair in a dataset.
}
\details{
Calculated using  \code{\link[DescTools]{KendallTauB}}. Assumes factor levels are in the given order.
NAs are automatically handled by pairwise omit.
}
\examples{
 d <- data.frame(x=rnorm(20), 
                 y=factor(sample(3,20, replace=TRUE)), 
                 z=factor(sample(2,20, replace=TRUE)))
 pair_tauB(d)
}
