% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_tauC}
\alias{pair_tauC}
\title{Stuarts's tau  C for association between ordinal factors.}
\usage{
pair_tauC(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{ignored. Pairwise complete observations are used automatically.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with factor pairs, or NULL if there are not at least two factor variables
}
\description{
Calculates Stuarts's tau C every  factor variable pair in a dataset.
}
\details{
Calculated using  \code{\link[DescTools]{StuartTauC}}. Assumes factor levels are in the given order.
NAs are automatically handled by pairwise omit.
}
\examples{
 d <- data.frame(x=rnorm(20), 
                 y=factor(sample(3,20, replace=TRUE)), 
                 z=factor(sample(2,20, replace=TRUE)))
 pair_tauC(d)
}
