% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\alias{pairwise.matrix}
\alias{pairwise.data.frame}
\title{A generic function to create a data structure for every variable pair in a dataset}
\usage{
pairwise(data, score = "?", pair_type = "?")

\method{pairwise}{matrix}(data, score = "?", pair_type = "?")

\method{pairwise}{data.frame}(data, score = NA_character_, pair_type = NA_character_)
}
\arguments{
\item{data}{A dataframe.}

\item{score}{a character string indicating the value of association.}

\item{pair_type}{a character string specifying the type of variable pair.}
}
\value{
A tbl_df of class \code{pairwise} for pairs of variables with a column \code{value} for the score value,
\code{score} for a type of association value and \code{pair_type} for the type of variable pair.
}
\description{
Creates a data structure for every variable pair in a dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{pairwise(matrix)}: pairwise method

\item \code{pairwise(data.frame)}: pairwise method

}}
\examples{
pairwise(cor(iris[,1:4]), score="pearson")
pairwise(iris)
}
