% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise.R
\name{plot.pairwise}
\alias{plot.pairwise}
\title{Plot method for class \code{pairwise}.}
\usage{
\method{plot}{pairwise}(x, type = c("matrix", "linear"), ...)
}
\arguments{
\item{x}{An object of class \code{pairwise}}

\item{...}{further arguments to \code{plot_pairwise} or \code{plot_pairwise_linear}}

\item{tyoe}{If "matrix", calls \code{plot_pairwise}, if "linear" calls \code{plot_pairwise_linear}}
}
\value{
a plot
}
\description{
Plot method for class \code{pairwise}.
}
\examples{
plot(pair_scores(iris))
}
