% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_db.R
\name{read_shapefiles}
\alias{read_shapefiles}
\title{Read shapefiles from a directory}
\usage{
read_shapefiles(dir_path)
}
\arguments{
\item{dir_path}{The directory containing the shapefile (.shp).}
}
\value{
A simple feature collection.
}
\description{
Read shapefiles from a directory
}
