% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullseye-package.R
\docType{package}
\name{bullseye-package}
\alias{bullseye}
\alias{bullseye-package}
\title{bullseye: Visualising Multiple Pairwise Variable Correlations and Other Scores}
\description{
We provide a tidy data structure and visualisations for multiple or grouped variable correlations, general association measures scagnostics and other pairwise scores suitable for numerical, ordinal and nominal variables. Supported measures include distance correlation, maximal information, ace correlation, Kendall's tau, and polychoric corrrelation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cbhurley.github.io/bullseye/}
  \item \url{https://github.com/cbhurley/bullseye}
}

}
\author{
\strong{Maintainer}: Catherine Hurley \email{catherine.hurley@mu.ie}

Authors:
\itemize{
  \item Amit Chinwan \email{amit.chinwan.2019@mumail.ie}
}

}
\keyword{internal}
