% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_scores.R
\name{pair_control}
\alias{pair_control}
\title{Default scores calculated  by \code{pairwise_scores}}
\usage{
pair_control(
  nn = "pair_cor",
  oo = "pair_polychor",
  ff = "pair_cancor",
  fn = "pair_cancor",
  nnargs = NULL,
  ooargs = NULL,
  ffargs = NULL,
  fnargs = NULL
)
}
\arguments{
\item{nn}{function for numeric pairs of variables, should return object of class \code{pairwise}. Use NULL to ignore numeric pairs.}

\item{oo}{function for ordered factor pairs of variables, should return object of class \code{pairwise}. Use NULL to ignore ordered factor pairs.}

\item{ff}{function for factor pairs of variables (not ordered), should return object of class \code{pairwise}. Use NULL to ignore factor-factor pairs.}

\item{fn}{function for factor-numeric pairs of variables, should return object of class \code{pairwise}. Use NULL to ignore factor-numeric pairs.}

\item{nnargs}{other arguments for the nn function}

\item{ooargs}{other arguments for the oo function}

\item{ffargs}{other arguments for the ff function}

\item{fnargs}{other arguments for the fn function}
}
\value{
list
}
\description{
Gives a list specifying the function to be used for two numeric (nn) variables, two factors (ff), two ordinals (oo)
and for a factor-numeric pair (fn).
}
