% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_polyserial}
\alias{pair_polyserial}
\title{Polyserial correlation}
\usage{
pair_polyserial(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{ignored. Pairwise complete observations are used automatically.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with polyserial correlation for factor-numeric pairs,
or NULL if there are not at least one such pair.
}
\description{
Calculates Polyserial correlation using  for every factor-numeric variable pair in a dataset.
}
\details{
The polyserial correlation is calculated using the \code{\link[polycor]{polyserial}} function from the
\code{polycor} package, and assumes factor levels are in the given order. NAs are automatically handled by pairwise omit.
}
\examples{
pair_polyserial(iris)
}
