% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_methods.R
\name{pair_uncertainty}
\alias{pair_uncertainty}
\title{Uncertainty coefficient for association between factors.}
\usage{
pair_uncertainty(d, handle.na = TRUE, ...)
}
\arguments{
\item{d}{A dataframe}

\item{handle.na}{ignored. Pairwise complete observations are used automatically.}

\item{...}{other arguments}
}
\value{
A tibble of class \code{pairwise} with every factor variable pair and uncertainty coefficient value,
or NULL if there are not at least two factor variables
}
\description{
Calculates uncertainty coefficient for every factor variable pair in a dataset.
}
\details{
The Uncertainty coefficient is calculated using \code{\link[DescTools]{UncertCoef}} function from the
\code{DescTools} package.
}
\examples{
 pair_uncertainty(iris)
}
